feat: 以 App Router RSC 遷移 recipe-list（資料抓取搬到 Server Component）
- page.tsx 改為 async Server Component：讀 searchParams（q/type/page）於 server 端呼叫 searchRecipesServer 抓取
- 排序/分頁改由 client 以 router.push 更新 URL → 觸發 Server Component 重新抓取（移除原 client loadRecipesByQuery + useEffect）
- next/head + PageSEO → generateMetadata（動態 SEO；title 用 absolute 避免與 root template 重複）
- getStaticProps 移除；searchRecipesServer 的 fetch 加 next.revalidate=3600（資料快取，對應原 revalidate）
- 頁面由 static+client-fetch 轉為 ƒ dynamic（server-rendered on demand）
- 驗證：next build 綠燈
